Implemented Skeleton for process, starter
Currently has only one response for every scenario.

I'm not sure how to implement the recommendation loop. Should we use a class variable (boolean)? If it is true and user inputs "yes", we give a recommendation. If the user inputs "no", the variable is changed to false, we print some response reset the recommendations vector. If something other than yes or no is given as input, we ask them to input either yes or no.

This would just require a conditional.