Fixed a bug in process, but titles are now printed with brackets
For loop made it so that after receiving the first yes, the final recommendation would be printed. So, only a maximum of two recommendations could be printed, the first and the last.

New code works, but it prints the title inside of square brackets with unnecessary infromation, so it needs a minor fix.

New code has only been applied to starte so it need to be added to creative